fix: the server can now return a valid token and refresh token on /refresh-token/new route

+++ src/main/java/com/keypass/server/token/RefreshTokenController.java
@@ -3,7 +3,6 @@ package com.keypass.server.token;
 import com.keypass.server.account.Account;
 import com.keypass.server.account.AccountService;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -13,6 +12,9 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 
+import java.time.LocalDateTime;
+import java.util.HashMap;
+
 @RestController
 @RequestMapping(value = "/refresh-token", produces = "application/json")
 @Tag(name = "Refresh Token", description = "Create and manage refresh tokens")
@@ -20,32 +22,42 @@ import lombok.RequiredArgsConstructor;
 public class RefreshTokenController {
   private final RefreshTokenService refreshTokenService;
   private final AccountService accountService;
-  @Autowired
   private final JwtService jwtService;
 
   @Operation(summary = "Create a new refresh token")
   @ApiResponses(value = {
       @ApiResponse(responseCode = "200", description = "Refresh token created"),
   })
-  @PostMapping("/new")
-  @ResponseBody
-  public ResponseEntity<Object> create(@RequestBody String refreshToken) {
-    RefreshToken storedRefreshToken = refreshTokenService.getRefreshTokenByHash(refreshToken.trim());
+  @PostMapping("/new/{userId}")
+  public ResponseEntity<Object> create(@PathVariable("userId") String userId, @RequestBody RefreshTokenRequestDto refreshTokenRequestDto) {
+    Account account = accountService.getAccountById(userId);
+    if (account == null) {
+      return ResponseEntity.badRequest().body("This refresh token does not belong to you");
+    }
+
+    RefreshToken storedRefreshToken = refreshTokenService.findRefreshTokenByUserId(account.getId().toString());
     if (storedRefreshToken == null) {
       return ResponseEntity.badRequest().body("Invalid refresh token");
     }
 
-    System.out.println(storedRefreshToken);
+    if(refreshTokenRequestDto.refreshToken().length() != storedRefreshToken.getToken().length()){
+      return ResponseEntity.badRequest().body("Size mismatch");
+    }
 
     refreshTokenService.deleteRefreshToken(storedRefreshToken.getId());
 
-    Account account = accountService.getAccountById(storedRefreshToken.getAccount().getId().toString());
-    if (account == null) {
-      return ResponseEntity.badRequest().body("Your refresh token does not belong to you");
-    }
+    HashMap<String, Object> keys =  jwtService.generateTokensWithoutAuth(account);
 
-    Object tokens = jwtService.generateTokensWithoutAuth(account);
+    RefreshToken databaseTokenGenerated = RefreshToken.builder()
+            .token(keys.get("refreshToken").toString())
+            .account(account)
+            .expiresIn(LocalDateTime.now().plusSeconds(1296000L))
+            .build();
 
-    return ResponseEntity.ok().body(tokens);
+    System.out.println(databaseTokenGenerated);
+
+    refreshTokenService.save(databaseTokenGenerated);
+
+    return ResponseEntity.ok().body(keys);
   }
 }
